# Conflicts:
#	Operadores.py

+++ Eleitor1.py
@@ -1,4 +1,4 @@
-idade = float(input('Entre com sua idade: '))
+idade = int(input('Entre com sua idade: '))
 if idade >=18:
     print('Maioridade. Voto obrigatório.')
 else:
